# meet-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Api for meet application.",
        "title": "meet API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "paths": {
...
        "/v1/chili-bank/account-types": {
            "get": {
                "description": "Get accounts types for chili banks",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chili Bank"
                ],
                "summary": "Get Account Types",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.ApiResponse-array_string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/v1/chili-bank/banks": {
            "get": {
                "description": "Get Chili Banks",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chili Bank"
                ],
                "summary": "Get Banks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.ApiResponse-array_string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/v1/chili-bank/{modelNickname}/accounts": {
            "get": {
                "description": "Get Model Chili Banks Accounts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chili Bank"
                ],
                "summary": "Get Model Accounts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "model nickname",
                        "name": "modelNickname",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.ApiResponse-array_dto_ChiliBankAccountDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "post": {
                "description": "Insert or update a chili bank account for the model",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chili Bank"
                ],
                "summary": "Save Model Account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "model nickname",
                        "name": "modelNickname",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The chili bank account data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ChiliBankAccountDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.ApiResponse-dto_ChiliBankAccountDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/v1/chili-bank/{modelNickname}/{accountId}": {
            "delete": {
                "description": "Delete a chili bank account of the model",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chili Bank"
                ],
                "summary": "Delete Model Account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "model nickname",
                        "name": "modelNickname",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "chili bank account id",
                        "name": "accountId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The chili bank account data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ChiliBankAccountDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.ApiResponse-string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.ChiliBankAccountDTO": {
            "type": "object",
            "required": [
                "accountNumber",
                "accountType",
                "bankName",
                "holderName",
                "rut"
            ],
            "properties": {
                "accountNumber": {
                    "type": "integer",
                    "maximum": 30
                },
                "accountType": {
                    "type": "string",
                    "maxLength": 50
                },
                "bankName": {
                    "type": "string",
                    "maxLength": 255
                },
                "holderName": {
                    "type": "string",
                    "maxLength": 255
                },
                "id": {
                    "type": "string",
                    "maxLength": 255
                },
                "rut": {
                    "type": "integer",
                    "maximum": 10
                }
            }
        },
...
        "rest.ApiErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "rest.ApiResponse-array_dto_ChiliBankAccountDTO": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/rest.ApiErrorDetail"
                },
                "payload": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ChiliBankAccountDTO"
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "rest.ApiResponse-array_string": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/rest.ApiErrorDetail"
                },
                "payload": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "rest.ApiResponse-dto_ChiliBankAccountDTO": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/rest.ApiErrorDetail"
                },
                "payload": {
                    "$ref": "#/definitions/dto.ChiliBankAccountDTO"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "rest.ApiResponse-string": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/rest.ApiErrorDetail"
                },
                "payload": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}